split config module into context and provider
