+++ deployments/stitch-frontend/index.html
@@ -21,7 +21,11 @@
         align-items: center;
         justify-content: center;
         gap: 0.75rem;
-        font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI",
+        font-family:
+          system-ui,
+          -apple-system,
+          BlinkMacSystemFont,
+          "Segoe UI",
           sans-serif;
         color: #475569;
         background: #fff;
