+++ deployments/stitch-frontend/src/components/ColophonPanel.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 import useBackendDiagnostics from "../hooks/useBackendDiagnostics";
-import { useConfig } from "../config/context";
+import { useConfig } from "../config/useConfig";
 
 function getConnectionInfo() {
   const nav = navigator;
