+++ deployments/stitch-frontend/src/config/context-provider.jsx
@@ -1,0 +1,11 @@
+import { ConfigContext } from "./context";
+
+export function ConfigProvider({ config, children }) {
+  if (!config || typeof config !== "object") {
+    throw new Error("ConfigProvider requires a valid config object.");
+  }
+
+  return (
+    <ConfigContext.Provider value={config}>{children}</ConfigContext.Provider>
+  );
+}
