+++ deployments/stitch-frontend/src/config/context.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const ConfigContext = createContext(null);
