+++ deployments/stitch-frontend/src/test/utils.jsx
@@ -2,7 +2,7 @@ import { vi } from "vitest";
 import { render } from "@testing-library/react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { MemoryRouter } from "react-router-dom";
-import { ConfigProvider } from "../config/context";
+import { ConfigProvider } from "../config/context-provider";
 import { getConfig } from "../config/env";
 
 /**
